Add ignore action to cancelled check
