+++ api/checks/summaries/pending.go
@@ -24,15 +24,6 @@ func (p Pending) GetSummary() (string, error) {
 	return compile(&p, "pending.md")
 }
 
-// CancelAction is an action that can be taken to cancel a pending check run.
-func CancelAction() *github.CheckRunAction {
-	return &github.CheckRunAction{
-		Identifier:  "cancel",
-		Label:       "Cancel",
-		Description: "Cancel this pending check run",
-	}
-}
-
 // GetActions returns the actions that can be taken by the user.
 func (p Pending) GetActions() []*github.CheckRunAction {
 	return []*github.CheckRunAction{
